weapon_app/requirements.txt
flask==3.1.3
ultralytics==8.4.33
deepface==0.0.99
tf-keras==2.21.0
tensorflow==2.21.0
opencv-python==4.13.0.92
numpy==2.2.6
pandas==2.3.3
Pillow==12.2.0
PyYAML==6.0.3
requests==2.33.1
tqdm==4.67.3
Werkzeug==3.1.8
